Moved some utility functions to won-utils.
