+++ webofneeds/won-owner-webapp/src/main/webapp/app/selectors.js
@@ -279,22 +279,5 @@ export const selectLastUpdatedPerConnection = createSelector(
     )
 );
 
-export function selectTimestamp(event, connectionUri) {
-    if(event.get('hasReceiver') === connectionUri) {
-        return event.get('hasReceivedTimestamp');
-    } else if(event.get('hasSender') === connectionUri) {
-        return event.get('hasSentTimestamp');
-    } else {
-        return undefined;
-    }
-};
-
-export function selectConnectionUris(need) {
-   return need
-       .getIn(['won:hasConnections', 'rdfs:member'])
-       .map(c => c.get('@id'));
-}
-
-
 window.selectAllByConnections4dbg = selectAllByConnections;
 window.allByConnection4db = allByConnection;
